+++ packages/amplication-data-service-generator/src/server/resource/dto/create-field-class-property.ts
@@ -39,8 +39,8 @@ import {
   SCALAR_FILTER_TO_MODULE_AND_TYPE,
 } from "./filters.util";
 import { createGraphQLFieldDecorator } from "./graphql-field-decorator";
-import { createCreateNestedManyWithoutInputID } from "./nested-input-dto/create-create-nested-many-without-input";
-import { createUpdateManyWithoutInputID } from "./nested-input-dto/create-update-many-without-input";
+import { createCreateNestedManyWithoutInputID } from "./nested-input-dto/create-nested";
+import { createUpdateManyWithoutInputID } from "./nested-input-dto/update-nested";
 import { JSON_VALUE_ID } from "./type-fest.util";
 import { createEntityListRelationFilterID } from "./graphql/entity-list-relation-filter/create-entity-list-relation-filter";
